@@ -1,8 +1,11 @@
 {
-  "name": "rest-express",
+  "name": "sentinelscope",
   "version": "1.0.0",
   "type": "module",
   "license": "MIT",
+  "engines": {
+    "node": ">=20"
+  },
   "scripts": {
     "dev": "dotenv -e .env -- tsx server/index.ts",
     "dev:bg": "nohup npm run dev >/tmp/sentinel_dev.log 2>&1 & echo $!",
@@ -11,7 +14,7 @@
     "start:bg": "nohup env PORT=3001 NODE_ENV=production node dist/index.js >/tmp/sentinel_prod.log 2>&1 & echo $!",
     "start:logs": "tail -f /tmp/sentinel_prod.log",
     "start:restart": "npm run dev:stop && npm run start:bg",
-    "dev:restart": "npm run start:restart",
+    "dev:restart": "npm run dev:stop && npm run dev:bg",
     "dev:local": "PORT=3001 JWT_SECRET=devsecret ALLOW_LEGACY_X_USER_ID=true DISABLE_RATE_LIMIT=1 npm run dev",
     "dev:pg": "cp .env.local.postgres .env && docker compose up -d db && npm run db:migrate && npm run dev",
     "dev:auto": "bash ./scripts/dev-auto.sh",
@@ -118,9 +121,9 @@
     "passport-local": "^1.0.0",
     "pg": "^8.11.3",
     "qrcode": "^1.5.4",
-    "react": "^18.3.1",
+    "react": "^19.2.1",
     "react-day-picker": "^8.10.1",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.2.1",
     "react-hook-form": "^7.55.0",
     "react-i18next": "^16.2.3",
     "react-icons": "^5.4.0",
@@ -149,8 +152,9 @@
     "@types/passport": "^1.0.16",
     "@types/passport-local": "^1.0.38",
     "@types/qrcode": "^1.5.6",
-    "@types/react": "^18.3.11",
-    "@types/react-dom": "^18.3.1",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
+    "@types/supertest": "^6.0.3",
     "@types/ws": "^8.5.13",
     "@vitejs/plugin-react": "^4.7.0",
     "@vitest/coverage-v8": "^2.1.9",
