@@ -4,6 +4,7 @@ import {
   AuthError,
   onAuthStateChanged,
   signInWithPopup,
+  signInWithRedirect,
   signInWithEmailAndPassword,
   createUserWithEmailAndPassword,
   sendPasswordResetEmail,
@@ -27,10 +28,12 @@ interface AuthContextType {
   mfaHint: MultiFactorInfo | null;
   loading: boolean;
   signInWithGoogle: () => Promise<void>;
+  signInWithGoogleRedirect: () => Promise<void>;
   signInWithEmail: (email: string, password: string) => Promise<void>;
   signUpWithEmail: (email: string, password: string) => Promise<void>;
   sendPasswordReset: (email: string) => Promise<boolean>;
   deleteAccount: (password: string) => Promise<boolean>;
+  sendMfaCode: () => Promise<string>;
   verifyMfaCode: (code: string) => Promise<void>;
   clearMfa: () => void;
   signOut: () => Promise<void>;
@@ -42,8 +45,9 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
   const [user, setUser] = useState<FirebaseUser | null>(null);
   const [isAdmin, setIsAdmin] = useState(false);
   const [loading, setLoading] = useState(true);
-    const [mfaResolver, setMfaResolver] = useState<MultiFactorResolver | null>(null);
-    const [mfaHint, setMfaHint] = useState<MultiFactorInfo | null>(null);
+  const [mfaResolver, setMfaResolver] = useState<MultiFactorResolver | null>(null);
+  const [mfaHint, setMfaHint] = useState<MultiFactorInfo | null>(null);
+  const [verificationId, setVerificationId] = useState<string>('');
   const { toast } = useToast();
 
   useEffect(() => {
@@ -59,13 +63,24 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
       });
     });
 
-    // Listen for auth state changes
-    const unsubscribe = onAuthStateChanged(auth, async (user) => {
+    let authResolved = false;
+
+    const handleAuthChange = async (user: FirebaseUser | null) => {
+      if (authResolved && !user) return; // Prevent overwriting if already resolved, unless user is signing out/changing
+      authResolved = true;
+
       // Dev auth override check
-      const isDev = import.meta.env.DEV || import.meta.env.VITE_DISABLE_FIREBASE_AUTH === 'true';
-      const devUserId = typeof localStorage !== 'undefined' ? localStorage.getItem('devUserId') : null;
+      // Allow dev auth if we are in dev mode OR if the flag is explicitly set
+      const isDevMode = import.meta.env.DEV || import.meta.env.VITE_DISABLE_FIREBASE_AUTH === 'true';
+      let devUserId = typeof localStorage !== 'undefined' ? localStorage.getItem('devUserId') : null;
       
-      if (isDev && devUserId && !user) {
+      // Default to 'demo' in dev mode if not set, to match queryClient behavior and ensure access
+      if (isDevMode && !devUserId) {
+        devUserId = 'demo';
+      }
+      
+      // Only use mock user if we are in dev mode AND a devUserId is set in localStorage
+      if (isDevMode && devUserId && !user) {
         // Mock user object for dev mode
         const mockUser = {
           uid: devUserId,
@@ -87,22 +102,35 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
           providerId: 'dev',
         } as unknown as FirebaseUser;
         
+        console.log('[Auth] Using dev mock user:', devUserId);
         setUser(mockUser);
         // Also sync the dev user
         try {
+          // Use a short timeout for the dev sync to prevent blocking
+          const controller = new AbortController();
+          const timeoutId = setTimeout(() => controller.abort(), 2000);
+          
           await fetch('/api/auth/user', {
             method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
+            headers: { 
+              'Content-Type': 'application/json',
+              'x-user-id': mockUser.uid
+            },
             body: JSON.stringify({
               id: mockUser.uid,
               email: mockUser.email,
               displayName: mockUser.displayName,
               photoURL: mockUser.photoURL,
             }),
+            signal: controller.signal
           });
+          clearTimeout(timeoutId);
           setIsAdmin(true); // Dev users are admins by default
-        } catch (e) {
-          console.error('Failed to sync dev user', e);
+        } catch (e: any) {
+           // Ignore abort errors or log them gently
+           if (e.name !== 'AbortError') {
+             console.error('Failed to sync dev user', e);
+           }
         }
         setLoading(false);
         return;
@@ -112,10 +140,14 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
       
       if (user) {
         try {
+          const token = await user.getIdToken();
           // Sync user with backend
           const createdResp = await fetch('/api/auth/user', {
             method: 'POST',
-            headers: { 'Content-Type': 'application/json' },
+            headers: { 
+              'Content-Type': 'application/json',
+              'Authorization': `Bearer ${token}`
+            },
             body: JSON.stringify({
               id: user.uid,
               email: user.email || '',
@@ -147,9 +179,23 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
       }
       
       setLoading(false);
-    });
+    };
 
-    return unsubscribe;
+    // Listen for auth state changes
+    const unsubscribe = onAuthStateChanged(auth, handleAuthChange);
+
+    // Fallback timeout: If Firebase takes too long (e.g. network block), force check dev mode
+    const fallbackTimeout = setTimeout(() => {
+      if (!authResolved) {
+        console.warn('[Auth] Firebase auth timed out, forcing state check...');
+        handleAuthChange(auth.currentUser);
+      }
+    }, 2500);
+
+    return () => {
+      clearTimeout(fallbackTimeout);
+      unsubscribe();
+    }
   }, [toast]);
 
   const signInWithGoogle = async () => {
@@ -162,12 +208,40 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
       });
     } catch (error: any) {
       console.error('Sign in error:', error);
+      
+      let description = error.message || "Failed to sign in. Please try again.";
+      
+      if (error.code === 'auth/popup-blocked') {
+        description = "Popup was blocked by your browser. Please allow popups for this site.";
+      } else if (error.code === 'auth/popup-closed-by-user') {
+        description = "Sign-in popup was closed before completion.";
+      } else if (error.code === 'auth/unauthorized-domain') {
+        description = "This domain (localhost) is not authorized in Firebase Console. Please add it to Authorized Domains.";
+      } else if (error.code === 'auth/operation-not-allowed') {
+        description = "Google Sign-In is not enabled in Firebase Console.";
+      }
+
       toast({
         title: "Authentication Failed",
-        description: error.message || "Failed to sign in. Please try again.",
+        description,
         variant: "destructive",
       });
     } finally {
+      setLoading(false);
+    }
+  };
+
+  const signInWithGoogleRedirect = async () => {
+    try {
+      setLoading(true);
+      await signInWithRedirect(auth, googleProvider);
+    } catch (error: any) {
+      console.error('Sign in redirect error:', error);
+      toast({
+        title: "Authentication Failed",
+        description: error.message || "Failed to initiate sign in.",
+        variant: "destructive",
+      });
       setLoading(false);
     }
   };
@@ -247,13 +321,41 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
   };
 
 
+  const sendMfaCode = async () => {
+    if (!mfaResolver || !mfaHint) {
+      throw new Error('No MFA resolver or hint available');
+    }
+    
+    try {
+      const phoneAuthProvider = new PhoneAuthProvider(auth);
+      const verId = await phoneAuthProvider.verifyPhoneNumber({
+        multiFactorHint: mfaHint,
+        session: mfaResolver.session
+      }, window.recaptchaVerifier);
+      
+      setVerificationId(verId);
+      return verId;
+    } catch (error: any) {
+      console.error('Error sending MFA code:', error);
+      toast({
+        title: "MFA Error",
+        description: error.message || "Failed to send verification code.",
+        variant: "destructive",
+      });
+      throw error;
+    }
+  };
+
   const verifyMfaCode = async (code: string) => {
     if (!mfaResolver) return;
     try {
       setLoading(true);
-  // TODO: Replace 'verificationId' with the actual verification ID from your MFA flow
-  const verificationId = '';
-  const phoneAuthCredential = PhoneAuthProvider.credential(verificationId, code);
+      
+      if (!verificationId) {
+        throw new Error('Verification ID not found. Please request a code first.');
+      }
+
+      const phoneAuthCredential = PhoneAuthProvider.credential(verificationId, code);
       const multiFactorAssertion = PhoneMultiFactorGenerator.assertion(phoneAuthCredential);
       await mfaResolver.resolveSignIn(multiFactorAssertion);
       clearMfa();
@@ -267,6 +369,7 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
   const clearMfa = () => {
     setMfaResolver(null);
     setMfaHint(null);
+    setVerificationId('');
   };
 
   const signOut = async () => {
@@ -290,17 +393,20 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
     <AuthContext.Provider value={{ 
       user,
       isAdmin,
-      loading, 
       mfaResolver,
       mfaHint,
-      signInWithGoogle, 
+      loading,
+      signInWithGoogle,
+      signInWithGoogleRedirect,
       signInWithEmail,
       signUpWithEmail,
       sendPasswordReset,
       deleteAccount,
+      sendMfaCode,
       verifyMfaCode,
       clearMfa,
-      signOut }}>
+      signOut,
+    }}>
       {children}
     </AuthContext.Provider>
   );
